fix adaptive card styling
